Add session settings and security

+++ E-Market-App/Controllers/AdminCategoriesController.cs
@@ -1,4 +1,5 @@
-﻿using EMarketApp.Core.Application.Interfaces.Services;
+﻿using E_Market_App.Middlewares;
+using EMarketApp.Core.Application.Interfaces.Services;
 using EMarketApp.Core.Application.ViewModels;
 using Microsoft.AspNetCore.Mvc;
 
@@ -7,20 +8,32 @@ namespace Pokedex_App.Controllers
     public class AdminCategoriesController : Controller
     {
         private readonly ICategoryService _categoryService;
+        private readonly ValidateUserSession _validateUserSession;
 
-        public AdminCategoriesController(ICategoryService categoryService)
+        public AdminCategoriesController(ICategoryService categoryService, ValidateUserSession validateUserSession)
         {
             _categoryService = categoryService;
+            _validateUserSession = validateUserSession;
         }
 
         public async Task<IActionResult> Index()
         {
+            if (!_validateUserSession.HasUser())
+            {
+                return RedirectToRoute(new { controller = "User", action = "Index" });
+            }
+
             ViewBag.Page = "adminCategories";
             return View("AdminCategories", await _categoryService.GetAllViewModel());
         }
 
         public IActionResult Create()
         {
+            if (!_validateUserSession.HasUser())
+            {
+                return RedirectToRoute(new { controller = "User", action = "Index" });
+            }
+
             ViewBag.Page = "adminCategories";
             return View("SaveCategory", new SaveCategoryViewModel());
         }
@@ -28,6 +41,11 @@ namespace Pokedex_App.Controllers
         [HttpPost]
         public async Task<IActionResult> Create(SaveCategoryViewModel vm)
         {
+            if (!_validateUserSession.HasUser())
+            {
+                return RedirectToRoute(new { controller = "User", action = "Index" });
+            }
+
             if (!ModelState.IsValid)
             {
                 return View("SaveCategory", vm);
@@ -39,6 +57,11 @@ namespace Pokedex_App.Controllers
 
         public async Task<IActionResult> Edit(int id)
         {
+            if (!_validateUserSession.HasUser())
+            {
+                return RedirectToRoute(new { controller = "User", action = "Index" });
+            }
+
             ViewBag.Page = "adminCategories";
             return View("SaveCategory", await _categoryService.GetSaveViewModelById(id));
         }
@@ -46,6 +69,11 @@ namespace Pokedex_App.Controllers
         [HttpPost]
         public async Task<IActionResult> Edit(SaveCategoryViewModel vm)
         {
+            if (!_validateUserSession.HasUser())
+            {
+                return RedirectToRoute(new { controller = "User", action = "Index" });
+            }
+
             if (!ModelState.IsValid)
             {
                 return View("SaveCategory", vm);
@@ -56,6 +84,11 @@ namespace Pokedex_App.Controllers
         }
         public async Task<IActionResult> Delete(int id)
         {
+            if (!_validateUserSession.HasUser())
+            {
+                return RedirectToRoute(new { controller = "User", action = "Index" });
+            }
+
             ViewBag.Page = "adminCategories";
             return View("DeleteCategory", await _categoryService.GetSaveViewModelById(id));
         }
@@ -63,6 +96,11 @@ namespace Pokedex_App.Controllers
         [HttpPost]
         public async Task<IActionResult> DeletePost(int id)
         {
+            if (!_validateUserSession.HasUser())
+            {
+                return RedirectToRoute(new { controller = "User", action = "Index" });
+            }
+
             await _categoryService.Delete(id);
             return RedirectToRoute(new { controller = "adminCategories", action = "Index" });
         }
